Improve policy examples
