+++ examples/mutation/image-prefix.ts
