+++ examples/validation/production-image-source.ts
